refactor(build): Extract webpub building into separate function
